requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
certifi=2021.5.30=py36haa95532_0
cffi=1.14.6=py36h2bbff1b_0
cycler=0.11.0=pypi_0
freetype=2.10.4=hd328e21_0
grep=0.3.2=pypi_0
icc_rt=2019.0.0=h0cc432a_1
intel-openmp=2021.4.0=haa95532_3556
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9d=h2bbff1b_0
kiwisolver=1.3.1=pypi_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
lz4-c=1.9.3=h2bbff1b_1
matplotlib=3.3.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36h196d8e1_0
mkl_fft=1.3.0=py36h46781fe_0
mkl_random=1.1.1=py36h47e9c7a_0
ninja=1.10.2=h559b2a2_2
numpy=1.19.2=py36hadc3359_0
numpy-base=1.19.2=py36ha3acd2a_0
olefile=0.46=py36_0
pandas=1.1.5=py36hd77b12b_0
pillow=8.3.1=py36h4fa10fc_0
pip=21.2.2=py36haa95532_0
pycparser=2.21=pyhd3eb1b0_0
pyparsing=3.0.6=pypi_0
python=3.6.5=h0c2934d_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.0.0=py3.6_cuda90_cudnn7_1
pytz=2021.3=pyhd3eb1b0_0
scikit-learn=0.24.2=py36hf11a4ad_1
scipy=1.5.2=py36h9439919_0
setuptools=58.0.4=py36haa95532_0
six=1.16.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h2bbff1b_0
torchvision=0.2.1=py_2
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.0=pyhd3eb1b0_1
wincertstore=0.2=py36h7fe50ca_0
xz=5.2.5=h62dcd97_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
